requirements.txt
# General dependencies
requests
PyYAML
ollama
Pygments
beautifulsoup4  # bs4 dependency
soupsieve  # To avoid warning with BeautifulSoup
httpx
python-docx>=1.1.2  # For processing Word documents
Pillow  # For image handling in formatting utils
openpyxl>=3.1.5  # For Excel workbook generation
et-xmlfile>=1.1.0  # openpyxl dependency
pytube>=15.0.0  # For YouTube video metadata and transcripts
yt-dlp>=2023.12.30
mutagen==1.47.0  # Audio file metadata handling, fixed version to avoid import issues
matplotlib>=3.7.0  # For charts in admin console
mermaid-py>=0.7.1  # For rendering Mermaid diagrams
ipython>=8.0.0  # For Mermaid magic functions in notebooks
defusedxml>=0.7.1  # Required for YouTube transcript node

# Optional audio libraries (uncomment to install)
librosa>=0.10.0  # For more accurate audio duration detection
soundfile>=0.12.1  # Alternative for audio file processing

# AI API dependencies
openai>=1.0.0  # For OpenAI API
anthropic>=0.8.0  # For Claude API
groq>=0.3.0  # For Groq API
google-genai>=0.1.0  # For Google Gemini API (new unified SDK)

# Core langchain dependencies
langchain
langchain-huggingface
langchain_core
langchain_community
sentence-transformers
pandas==1.5.3  # Changed to a more compatible version for Python 3.12 on Windows
numpy==1.26.4
transformers
typing_extensions
pydantic==2.9.2

# Additional dependencies for transformers
filelock
huggingface-hub<1.0,>=0.23.2
regex!=2019.12.17
safetensors>=0.4.1
tokenizers<0.21,>=0.20
tqdm

# Dependencies for faiss-cpu and others
faiss-cpu
urllib3<2.0.0

# Modules required for pydantic and httpx functionality
pydantic_core==2.23.4
annotated_types
sniffio
httpcore
h11

# Pandas dependencies
python-dateutil>=2.8.2
pytz>=2020.1
tzdata>=2022.7

# PyTorch is installed by setup.bat with the correct CPU/GPU wheel

# YouTube transcript dependencies
youtube-transcript-api>=0.6.2

# GUI dependencies
PyQt5>=5.15.11  # Required for OutlineWriterNode review window

pillow
scikit-learn==1.5.0  # Specific version for compatibility
scipy
fsspec
joblib
threadpoolctl
colorama
jinja2
networkx
sympy
mpmath<1.4,>=1.1.0
pypdf  # For PDF support
markupsafe  # Required by Jinja2

# Text-to-speech dependencies
gradio_client>=0.10.0
pydub

# File processing dependencies
watchdog  # For file system monitoring
pydub  # For audio processing
PyPDF2  # For PDF processing
quickchart-io  # For chart generation in ExportWord module
# Note: Mermaid diagrams are limited to ~784px width due to mermaid.ink API constraints
# SVG conversion libraries (cairosvg, svglib) have compatibility issues with Mermaid SVG output
